ALL_BUILDS
Well there's the problem. Little tidbit of code to sort out the tri-stated flags was wrecking the ALL_BUILDS flag. Since nothing is currently in slot #12 except for my stats drop down I just chopped the loop to stop a smidge earlier. So now this whole branch is working exactly like it's supposed to work.